Moved code related to serialization to its own package.
