+++ MazesAndMinotaurs/app/src/main/java/com/example/cis/mazeminotaurs/character/SaveAndLoadDialog.java
@@ -14,6 +14,7 @@ import android.widget.Button;
 import com.example.cis.mazeminotaurs.CharacterSheetFragment;
 import com.example.cis.mazeminotaurs.Portfolio;
 import com.example.cis.mazeminotaurs.R;
+import com.example.cis.mazeminotaurs.serialization.SaveAndLoadPerformer;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
